feat: add exponential backoff to API client
The API client now retries retrieving the serviceKey via
the initialization request in an exponential backoff loop.
The exponential algorithm utilizes an error factor to
spread out the retrials, so when multiple clients are started
at the same time they are unlikely to issue any request after
the initial one at the same time.